+++ pkg/driver/ydb/driver.go
@@ -35,12 +35,13 @@ func init() {
 }
 
 type Driver struct {
-	db       *sql.DB
-	nativeDB *ydbsdk.Driver
-	dialect  queries.Dialect
-	logger   *zap.Logger
-	sqlCfg   *stroppy.DriverConfig_SqlConfig
-	bulkSize int
+	db           *sql.DB
+	nativeDB     *ydbsdk.Driver
+	dialect      queries.Dialect
+	logger       *zap.Logger
+	sqlCfg       *stroppy.DriverConfig_SqlConfig
+	bulkSize     int
+	queryTimeout time.Duration
 }
 
 var _ driver.Driver = (*Driver)(nil)
@@ -88,12 +89,13 @@ func NewDriver(
 	}
 
 	return &Driver{
-		db:       db,
-		nativeDB: nativeDB,
-		dialect:  ydbDialect{},
-		logger:   lg,
-		sqlCfg:   sqlCfg,
-		bulkSize: bulkSize,
+		db:           db,
+		nativeDB:     nativeDB,
+		dialect:      ydbDialect{},
+		logger:       lg,
+		sqlCfg:       sqlCfg,
+		bulkSize:     bulkSize,
+		queryTimeout: opts.QueryTimeout,
 	}, nil
 }
 
@@ -194,7 +196,7 @@ func (d *Driver) Begin(ctx context.Context, isolation stroppy.TxIsolationLevel) 
 			return nil, err
 		}
 
-		return sqldriver.NewConnOnlyTx(conn, sqldriver.NewRows, d.dialect, d.logger, conn.Close), nil
+		return sqldriver.NewConnOnlyTx(conn, sqldriver.NewRows, d.dialect, d.logger, d.queryTimeout, conn.Close), nil
 	}
 
 	sqlTx, err := d.db.BeginTx(ctx, &sql.TxOptions{Isolation: sqldriver.IsolationToSQL(isolation)})
@@ -208,6 +210,7 @@ func (d *Driver) Begin(ctx context.Context, isolation stroppy.TxIsolationLevel) 
 		isolation,
 		d.dialect,
 		d.logger,
+		d.queryTimeout,
 	), nil
 }
 
@@ -216,7 +219,7 @@ func (d *Driver) RunQuery(
 	sqlStr string,
 	args map[string]any,
 ) (*driver.QueryResult, error) {
-	return sqldriver.RunQuery(ctx, d.db, sqldriver.NewRows, d.dialect, d.logger, sqlStr, args)
+	return sqldriver.RunQuery(ctx, d.db, sqldriver.NewRows, d.dialect, d.logger, sqlStr, args, d.queryTimeout)
 }
 
 func (d *Driver) Teardown(ctx context.Context) error {